Bump aiovantage version, move event handling to events.py
