+++ .pre-commit-config.yaml
@@ -15,7 +15,7 @@ repos:
     hooks:
       - id: mypy
         additional_dependencies:
-          - aiovantage==0.7.0
+          - aiovantage==0.8.0
           - voluptuous-stubs==0.1.1
           - homeassistant-stubs==2023.7.1
         args: []
