+++ custom_components/vantage/switch.py
@@ -4,7 +4,7 @@ import functools
 from typing import Any
 
 from aiovantage import Vantage
-from aiovantage.config_client.objects import GMem, Load
+from aiovantage.models import GMem, Load
 
 from homeassistant.components.switch import SwitchEntity
 from homeassistant.config_entries import ConfigEntry
