+++ requirements.txt
@@ -3,7 +3,7 @@ homeassistant==2023.7.1
 colorlog==6.7.0
 
 # Use a specific version of aiovantage
-aiovantage==0.7.0
+aiovantage==0.8.0
 
 # Use a local version of aiovantage
 # -e ../aiovantage
